refactor: deduplicate health threshold modules
