+++ src/core/health/__init__.py
@@ -1,42 +1,2 @@
-#!/usr/bin/env python3
-"""
-Health Threshold Management Package - Agent_Cellphone_V2
-
-Extracted and refactored health threshold management components.
-Part of the HealthThresholdManager refactoring for SRP compliance.
-
-Author: Agent-7 (Refactoring Specialist)
-License: MIT
-"""
-
-from .models import (
-    HealthMetricType,
-    HealthThreshold,
-    ThresholdOperation,
-    ValidationOperation,
-    ConfigurationChange
-)
-
-from .operations import HealthThresholdOperations
-from .validation import HealthThresholdValidation
-from .persistence import HealthThresholdPersistence
-from .defaults import HealthThresholdDefaults
-from .testing import HealthThresholdTesting
-from .monitoring import HealthThresholdMonitoring
-
-__all__ = [
-    # Models
-    'HealthMetricType',
-    'HealthThreshold',
-    'ThresholdOperation',
-    'ValidationOperation',
-    'ConfigurationChange',
-    
-    # Services
-    'HealthThresholdOperations',
-    'HealthThresholdValidation',
-    'HealthThresholdPersistence',
-    'HealthThresholdDefaults',
-    'HealthThresholdTesting',
-    'HealthThresholdMonitoring',
-]
+from core.health_base import *  # noqa: F401,F403
+from core.health_base import __all__ as __all__
