+++ src/core/health_base/__init__.py
@@ -1,0 +1,41 @@
+#!/usr/bin/env python3
+"""
+Health Threshold Management Package - Agent_Cellphone_V2
+
+Extracted and refactored health threshold management components.
+Part of the HealthThresholdManager refactoring for SRP compliance.
+
+Author: Agent-7 (Refactoring Specialist)
+License: MIT
+"""
+
+from .models import (
+    HealthMetricType,
+    HealthThreshold,
+    ThresholdOperation,
+    ValidationOperation,
+    ConfigurationChange,
+)
+
+from .operations import HealthThresholdOperations
+from .validation import HealthThresholdValidation
+from .persistence import HealthThresholdPersistence
+from .defaults import HealthThresholdDefaults
+from .testing import HealthThresholdTesting
+from .monitoring import HealthThresholdMonitoring
+
+__all__ = [
+    # Models
+    "HealthMetricType",
+    "HealthThreshold",
+    "ThresholdOperation",
+    "ValidationOperation",
+    "ConfigurationChange",
+    # Services
+    "HealthThresholdOperations",
+    "HealthThresholdValidation",
+    "HealthThresholdPersistence",
+    "HealthThresholdDefaults",
+    "HealthThresholdTesting",
+    "HealthThresholdMonitoring",
+]
